Fix schematic export bug
The version available by apt-get of Lepton EDA
forgets to declare the signals of NC (not connected) pins.
Perhaps, they plan to call sub circuits, using "open".
The work around is to force their name.

However, some additional work is needed to make the syntheses up and running
as I have to compile, in the workflow, the ghdl-yosys plug-in