Fix a FATAL bug
The NJW4191 can accept large power supply voltage.
However the ON/OFF voltage should not exceed 6V.

The new parts-list and schematic diagrams are going to be produced by the artifact.

Since this project is a fun demo project,
Since the charge pump has not been placed nor routed on the PCB,
The PCB is not going to be updated.